+++ frontend/src/composables/RESTAPI.ts
@@ -5,6 +5,7 @@ import {
   AuthenticationError,
   PermissionError,
   APIServerError,
+  CanceledRequestError,
   RequestError
 } from '@/composables/HTTPErrors'
 
@@ -56,6 +57,9 @@ export function useRESTAPI() {
         }
       } else if (error.request) {
         /* No reply from server */
+        //console.log(error)
+        if (error.code == "ERR_CANCELED")
+          throw new CanceledRequestError('Canceled request')
         throw new RequestError(`Request error: ${error.message}`)
       } else {
         /* Something else happening when setting up the request */